Refactor gli.io
